main.py: move check_risk to the language agnostic model and enable it
The ML and Research team have worked a lot on the Revert Risk Language
Agnostic model, it should support the Serobot's use case and return
better predictions than ORES. The multi-lingual model is still
under performing so the Research team discourages its use, for
the moment.

Update the check_risk function to use language agnostic, and use
it instead the checkORES one.

Issue: #5